Create The configuration panel
