+++ yaqbot/src/App.css
@@ -24,6 +24,8 @@
   --soft-green-color: #b7e7a3;
 
   --medium-grey-color: #888;
+
+  --soft-red-color: #cf2f2f;
 }
 
 #root {
