+++ yaqbot/src/components/molecules/DescriptionCard/DescriptionCard.module.css
@@ -7,7 +7,7 @@
 
 .description h1 {
   font-family: var(--header-title);
-  font-size: 1.25rem;
+  font-size: 0.9rem;
 }
 
 .description p {
