+++ yaqbot/src/components/molecules/PlantConfiguration/PlantConfiguration.module.css
@@ -1,0 +1,31 @@
+.configuration_board {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid var(--primary-darker-color);
+  border-radius: 0.5rem;
+  padding: 1rem;
+  gap: 0.5rem;
+  justify-content: space-between;
+  background-color: var(--soft-grey-color);
+}
+
+.configuration_board h1 {
+  font-family: var(--header-title);
+  font-size: 1.25rem;
+  text-align: left;
+}
+
+.action_btn {
+  background-color: var(--primary-white-color);
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  cursor: pointer;
+  border: 1px solid var(--primary-dark-color);
+}
+
+.delete_btn {
+  background-color: var(--soft-red-color);
+  color: var(--primary-white-color);
+  font-weight: bold;
+  border: none;
+}
