+++ yaqbot/src/components/molecules/ToggleOption/ToggleOption.module.css
@@ -1,6 +1,6 @@
 .option_body {
   display: flex;
   justify-content: space-between;
-  padding: 1rem;
+  padding: 0.5rem;
   align-items: center;
 }
